remote(wsl): Make shell and platform discovery more resilient to shell scripts
